+++ src/AssetsPlatform/AwsS3Adapter.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace servd\AssetStorage\volumeDriver;
+namespace servd\AssetStorage\AssetsPlatform;
 
 use League\Flysystem\AwsS3v3\AwsS3Adapter as OriginalAwsS3Adapter;
 use Aws\S3\S3Client;